requirements.txt
# NewsGPT Navigator — Python Dependencies

# LangGraph orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# LLM (Groq - free tier)
groq>=0.9.0

# Embeddings & vector store
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# News fetching
requests>=2.31.0

# Translation
deep-translator>=1.11.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Config
python-dotenv>=1.0.0

# Video generation pipeline
Pillow>=10.0.0
gTTS>=2.4.0
moviepy>=1.0.3

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
